add guards and profile page

+++ JS Back-End Exam - 27 June 2021/Shared Trip/controllers/tripController.js
@@ -5,7 +5,7 @@
 // if (Object.values(req.body).some(x => !x)) {
 //     throw new Error('All fields are required!')
 // }const { isGuest } = require('../middlewares/guard');
-const { getAll, createTrip, getById, editById, deleteById, buyCrypto, searchRefDATA, addUserToItem } = require('../services/itemServices');
+const { getAll, createTrip, getById, editById, deleteById, buyCrypto, searchRefDATA, addUserToItem, getByIdWithOwnerAndBuddies, joinTrip, editSeats } = require('../services/itemServices');
 const { parseError } = require('../util/parser');
 const { hasUser, isGuest } = require('../middlewares/guard');
 const { updateUser } = require('../services/userService');
@@ -34,14 +34,14 @@ tripController.get('/', async (req, res) => {
         }
 })
 
-tripController.get('/create', (req, res) => {
+tripController.get('/create',hasUser(), (req, res) => {
         res.render('trip-create', {
                 title: 'Offer trip',
                 user: req.user
         })
 })
 
-tripController.post('/create', async (req, res) => {
+tripController.post('/create',hasUser(), async (req, res) => {
 
         const trip = {
                 startPoint: req.body.startPoint,
@@ -71,78 +71,87 @@ tripController.post('/create', async (req, res) => {
 
 
 tripController.get('/details/:id', async (req, res) => {
-        const trip = await getById(req.params.id)
-        console.log(trip);
-        // //isOwner is for edit and delete functionality
-        // crypto.isOwner = crypto.owner.toString() == (req.user?._id)?.toString();
+        const trip = await getByIdWithOwnerAndBuddies(req.params.id)
 
-        // crypto.bayer = crypto.buyer.map(x => x.toString()).includes(req.user?._id.toString())
+
+        trip.isOwner = trip.owner._id.toString() == (req.user?._id)?.toString();
+        trip.hasSeats = trip.seats > 0 ? true : false
+
+
+        trip.hasBuddies = trip.buddies.map(x => x._id.toString()).includes(req.user?._id.toString())
+        const buddies = []
+        trip.buddies.map(x => buddies.push(x.email))
+
         res.render('trip-details', {
                 title: 'Details Trip',
                 user: req.user,
-                trip
+                trip,
+                buddies: buddies.join(', ')
         })
 });
 
-tripController.get('/edit/:id', async (req, res) => {
+tripController.get('/edit/:id',hasUser(), async (req, res) => {
         //TODO guard for Owner
-        const crypto = await getById(req.params.id)
-        const isOwner = crypto.owner.toString() == (req.user?._id)?.toString();
+        const trip = await getById(req.params.id)
+        const isOwner = trip.owner.toString() == (req.user?._id)?.toString();
         if (!isOwner) {
-                res.redirect('/auth/login')
+                res.redirect('/')
         }
 
-        res.render('edit', {
-                title: 'Edit Page',
+        res.render('trip-edit', {
+                title: 'Edit Trip',
                 user: req.user,
-                crypto,
+                trip,
         })
 })
 
-tripController.post('/edit/:id', async (req, res) => {
+tripController.post('/edit/:id',hasUser(), async (req, res) => {
         //TODO guard for Owner
-        const crypto = await getById(req.params.id)
-        const isOwner = crypto.owner.toString() == (req.user?._id)?.toString();
+        const trip = await getById(req.params.id)
+        const isOwner = trip.owner.toString() == (req.user?._id)?.toString();
         if (!isOwner) {
                 res.redirect('/auth/login')
         }
 
         try {
                 await editById(req.params.id, req.body)
-                res.redirect(`/crypto/details/${req.params.id}`)
+                res.redirect(`/catalog/details/${req.params.id}`)
         } catch (error) {
-                res.render('edit', {
-                        error: parseError(error),
-                        crypto,
+                res.render('trip-edit', {
+                        errors: parseError(error),
+                        title: 'Edit Trip',
+                        trip,
                         user: req.user
                 })
         }
 });
 
 
-tripController.get('/delete/:id', async (req, res) => {
-        const crypto = await getById(req.params.id)
-        const isOwner = crypto.owner.toString() == (req.user?._id)?.toString();
+tripController.get('/delete/:id',hasUser(), async (req, res) => {
+        const trip = await getById(req.params.id)
+        const isOwner = trip.owner.toString() == (req.user?._id)?.toString();
 
         if (!isOwner) {
-                return res.redirect(`/auth/login/`)
+                return res.redirect(`/`)
         }
         await deleteById(req.params.id)
-        res.redirect('/crypto/catalog')
+        res.redirect('/catalog')
 });
 
 
-tripController.get('/buy/:id', isGuest(), async (req, res) => {
-        const crypto = await getById(req.params.id)
-        if (crypto.owner.toString() != (req.user?._id)?.toString()
-                && crypto.buyer.map(x => x.toString()).includes((req.user?._id)?.toString()) == false) {
+tripController.get('/join/:id',hasUser(), async (req, res) => {
+        const trip = await getById(req.params.id)
+        await editSeats(req.params.id)
+
+        if (trip.owner.toString() != (req.user?._id)?.toString()
+                && trip.buddies.map(x => x.toString()).includes((req.user?._id)?.toString()) == false && trip.seats > 0) {
                 try {
-                        await buyCrypto(req.params.id, req.user._id);
-                        res.redirect(`/crypto/details/${req.params.id}`)
+                        await joinTrip(req.params.id, req.user._id);
+                        res.redirect(`/catalog/details/${req.params.id}`)
                 } catch (error) {
-                        res.render('catalog', {
-                                error: parseError(error),
-                                crypto,
+                        res.render('404', {
+                                errors: parseError(error),
+                                title: 'Error Page',
                                 user: req.user
                         })
                 }
